Fix build with Xcode clang.
The clang installation provided by Xcode includes two proxy executables
named 'gcc' and 'g++'.  These actually invoke clang/++ but in a posix-
compatible mode (after dealing with -D/U etc.).

However, what this means is that on macOS platforms, with Xcode installed
it looks as if 'g++' exists - but that g++ is only compatible with FSF
g++-4.2.1, so either one needs to test the GCC version as well as for its
existence before adding flags, or one needs to be explicit to use clang++.

This adds a clause to Makefile.in to cater for clang++ (a more general
solution could be as above).