cmake : Fix the build as a dependency of clang.
The existing cmake implementation was not working when libcody
was a dependency of a clang library (although it was building fine
when built as part of the LLVM 'all' target).  Update the cmake
recipe to be specific about stand-alone build, and to carry out
the configuration for in-tree build otherwise.